+++ protostar/test_runner/test_runner.py
@@ -8,7 +8,6 @@ from typing import List, Optional
 from starkware.starknet.services.api.contract_class import ContractClass
 from starkware.starkware_utils.error_handling import StarkException
 
-from protostar.commands.test.starkware.test_execution_state import TestExecutionState
 from protostar.commands.test.test_environment_exceptions import ReportedException
 from protostar.protostar_exception import ProtostarException
 from protostar.test_runner.environments.setup_execution_environment import (
@@ -20,6 +19,7 @@ from protostar.utils.compiler.pass_managers import (
 )
 from protostar.utils.starknet_compilation import CompilerConfig, StarknetCompiler
 
+from .starkware.test_execution_state import TestExecutionState
 from .test_case_runners.setup_case_runner import run_setup_case
 from .test_case_runners.test_case_runner_factory import TestCaseRunnerFactory
 from .test_config import TestConfig
