+++ tests/benchmarks/test_benchmark.py
@@ -10,8 +10,8 @@ from typing import List, Optional, Tuple
 import pytest
 from starkware.starknet.services.api.contract_class import ContractClass
 
-from protostar.commands.test.starkware.test_execution_state import TestExecutionState
 from protostar.test_runner import SharedTestsState, TestCollector, TestRunner
+from protostar.test_runner.starkware.test_execution_state import TestExecutionState
 from protostar.test_runner.test_config import TestConfig
 from protostar.test_runner.test_suite import TestCase, TestSuite
 from protostar.utils.compiler.pass_managers import ProtostarPassMangerFactory
